pyproject.toml
[project]
name = "simple-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Seoyeon Kim",email = "seoyeonkim0404@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "simple-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.0.0"
uvicorn = "^0.35.0"

